Initial commit to CMBackgroundChooser
This is a helper package for CMParts,
allowing the user to select an image
from the sdcard to use as a background
in various GUI elements.

This initial commit only supports the
notification dropdown background, but
it is easily extended to support other
GUI elements.

Original concept taken from Pedlar's
CMBootanimation package